python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quantqc"
version = "0.1.0"
description = "Single-cell mass spectrometry proteomics QC pipeline"
requires-python = ">=3.10"
dependencies = [
    "polars>=0.20",
    "pyarrow>=12.0",
    "numpy>=1.24",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "scikit-learn>=1.2",
    "biopython>=1.81",
    "pandas>=2.0",
]

[project.optional-dependencies]
umap = ["umap-learn>=0.5"]
cluster = ["leidenalg>=0.9", "python-igraph>=0.10"]
batch = ["combat>=0.3"]
xgboost = ["xgboost>=1.7"]
all = ["quantqc[umap,cluster,batch,xgboost]"]

[tool.hatch.build.targets.wheel]
packages = ["quantqc"]
